plan: move IdentEvalVisitor to plans package and use Getcdentvalue function.
